requirements.txt
customtkinter
PyJWT
